pyproject.toml
[tool.poetry]
name = "ecom"
version = "0.1.0"
description = ""
authors = ["Carlos Porta <cmaciasporta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlalchemy = "^2.0.13"
aiosqlite = "^0.19.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
result = "^0.10.0"
pyotp = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
httpx = "^0.24.0"
pytest-asyncio = "^0.21.0"
ruff = "^0.0.265"
black = "^23.3.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
